+++ .build
@@ -2,4 +2,4 @@ Build information, patch level is incremented at compilation.
 
 RELEASE	=	0
 VERSION	=	0
-PATCH	=	477
+PATCH	=	544
