+++ srcs/core/resolve.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/19 15:04:42 by arsciand          #+#    #+#             */
-/*   Updated: 2021/11/21 12:30:52 by arsciand         ###   ########.fr       */
+/*   Updated: 2021/11/22 20:14:32 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,9 +77,8 @@ uint8_t     resolve_target(t_traceroute *traceroute, char *target, int argc)
 
     /* Preparing getaddrinfo struct */
     hints.ai_flags      = AI_V4MAPPED | AI_ALL | AI_ADDRCONFIG;
-    hints.ai_family     = AF_UNSPEC;
+    hints.ai_family     = AF_INET; // AF_UNSPEC for both IPV4 and IPV6
     hints.ai_socktype   = SOCK_RAW;
-    // hints.ai_protocol   = IPPROTO_ICMP;
 
     if ((status = getaddrinfo(target, NULL, &hints, &res)) != SUCCESS)
     {
@@ -98,8 +97,8 @@ uint8_t     resolve_target(t_traceroute *traceroute, char *target, int argc)
             traceroute->conf.mode |= AF_INET;
             ((struct sockaddr_in *)&traceroute->target)->sin_addr.s_addr
                 = ((struct sockaddr_in *)res->ai_addr)->sin_addr.s_addr;
-            // ((struct sockaddr_in *)&traceroute->target)->sin_port
-            //     = htons(0);
+            ((struct sockaddr_in *)&traceroute->target)->sin_port
+                = htons(0);
             ((struct sockaddr_in *)&traceroute->target)->sin_family
                 = (sa_family_t)res->ai_family;
             if (inet_pton_handler(traceroute, target) != TRUE)
