+++ srcs/tools/handlers.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/19 15:32:23 by arsciand          #+#    #+#             */
-/*   Updated: 2021/11/21 20:02:34 by arsciand         ###   ########.fr       */
+/*   Updated: 2021/11/22 20:13:48 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void        getnameinfo_handler(t_traceroute *traceroute)
         if ((status = getnameinfo((struct sockaddr *)&sa, sizeof(sa), traceroute->buff_dns,
                     sizeof(traceroute->buff_dns), NULL, 0, NI_NAMEREQD)) != 0)
         {
-            if (status != EAI_NONAME && status != EAI_FAMILY)
+            if (status != EAI_NONAME)
                 getnameinfo_error_handler(traceroute, status);
         }
         inet_ntop_handler(traceroute, (uint32_t *)&sa.sin_addr);
